clean up calanderization PR
